Add favicons to layout template
The assets were generated based on `assets/favicon/phlex_icon.png`
which was cropped from `phlex_logo.png`. Because the source was
164x164 pixels, this omits larger-sized favicon declarations since
the upscaling looks poor.

Recreating the icon sizes using an SVG source would be a nice
enhancement, and allow for larger favicons to be included.

Icon sizes and declarations were generated using
https://realfavicongenerator.net/