Add phlex and phlexecute markdown code blocks
`phlex` code blocks will be `eval`ed in an execution context scoped to
the markdown file, and the code that was `eval`ed will be rendered as
a Ruby code block.

`phlexecute` code blocks will also be `eval`ed in the same execution
context, scoped to the markdown file, but the resulting output of the
`eval` will be rendered as an HTML code block.