adopt ”smart” quotes globally outside of code
This is a debatable change, since the double quotes appear in a title.  If you prefer to keep the style of double quote that originally appeared in the title on YouTube, then reject this PR.  If you want to use ”smart” quotes anyway, then this is the only place in the docs that I could find ”smart” quotes missing.

It sort of comes down to whether you consider the style of punctuation to be a purely stylistic choice, or whether you consider it to itself be content and therefore preserving it is necessary to accurately represent the original.  

Splitting hairs here, of course, because _everywhere_ else (outside of code) the documentation already had everything ”smartened”.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/phlex-ruby/phlex.fun?shareId=XXXX-XXXX-XXXX-XXXX).